feat: add rc_joins_remote_burst_count and rc_joins_remote_per_second (#477)
