+++ src/synapse/__init__.py
@@ -59,7 +59,17 @@ from .workload import (  # noqa: F401
     SYNAPSE_WORKER_CONFIG_PATH,
     ExecResult,
     WorkloadError,
-    create_mjolnir_config,
+    execute_migrate_config,
+    generate_mjolnir_config,
+    generate_nginx_config,
+    generate_worker_config,
+    get_environment,
+    get_media_store_path,
+    get_registration_shared_secret,
+    reset_instance,
+    validate_config,
+)
+from .workload_configuration import (  # noqa: F401
     disable_password_config,
     disable_room_list_search,
     enable_allow_public_rooms_over_federation,
@@ -70,6 +80,7 @@ from .workload import (  # noqa: F401
     enable_media,
     enable_media_retention,
     enable_metrics,
+    enable_rc_joins_remote_rate,
     enable_redis,
     enable_replication,
     enable_room_list_publication_rules,
@@ -79,12 +90,4 @@ from .workload import (  # noqa: F401
     enable_stale_devices_deletion,
     enable_stream_writers,
     enable_trusted_key_servers,
-    execute_migrate_config,
-    generate_nginx_config,
-    generate_worker_config,
-    get_environment,
-    get_media_store_path,
-    get_registration_shared_secret,
-    reset_instance,
-    validate_config,
 )
