+++ src/synapse/workload.py
@@ -216,6 +216,41 @@ def _exec(
         )
 
 
+def get_environment(charm_state: CharmState) -> typing.Dict[str, str]:
+    """Generate a environment dictionary from the charm configurations.
+
+    Args:
+        charm_state: Instance of CharmState.
+
+    Returns:
+        A dictionary representing the Synapse environment variables.
+    """
+    environment = {
+        "SYNAPSE_CONFIG_DIR": SYNAPSE_CONFIG_DIR,
+        "SYNAPSE_CONFIG_PATH": SYNAPSE_CONFIG_PATH,
+        "SYNAPSE_DATA_DIR": SYNAPSE_DATA_DIR,
+        "SYNAPSE_REPORT_STATS": f"{charm_state.synapse_config.report_stats}",
+        "SYNAPSE_SERVER_NAME": f"{charm_state.synapse_config.server_name}",
+        # TLS disabled so the listener is HTTP. HTTPS will be handled by Traefik.
+        # TODO verify support to HTTPS backend before changing this  # pylint: disable=fixme
+        "SYNAPSE_NO_TLS": str(True),
+        "LD_PRELOAD": "/usr/lib/x86_64-linux-gnu/libjemalloc.so.2",
+    }
+    datasource = charm_state.datasource
+    if datasource is not None:
+        environment["POSTGRES_DB"] = datasource["db"]
+        environment["POSTGRES_HOST"] = datasource["host"]
+        environment["POSTGRES_PORT"] = datasource["port"]
+        environment["POSTGRES_USER"] = datasource["user"]
+        environment["POSTGRES_PASSWORD"] = datasource["password"]
+    for proxy_variable in ("http_proxy", "https_proxy", "no_proxy"):
+        proxy_value = getattr(charm_state.proxy, proxy_variable)
+        if proxy_value:
+            environment[proxy_variable] = str(proxy_value)
+            environment[proxy_variable.upper()] = str(proxy_value)
+    return environment
+
+
 def execute_migrate_config(container: ops.Container, charm_state: CharmState) -> None:
     """Run the Synapse command migrate_config.
 
@@ -266,472 +301,6 @@ def validate_config(container: ops.Container) -> None:
         raise WorkloadError("Validate config failed, please check the logs")
 
 
-def enable_metrics(current_yaml: dict) -> None:
-    """Change the Synapse configuration to enable metrics.
-
-    Args:
-    current_yaml: current configuration.
-
-    Raises:
-        EnableMetricsError: something went wrong enabling metrics.
-    """
-    try:
-        metric_listener = {
-            "port": int(SYNAPSE_EXPORTER_PORT),
-            "type": "metrics",
-            "bind_addresses": ["::"],
-        }
-        current_yaml["listeners"].extend([metric_listener])
-        current_yaml["enable_metrics"] = True
-    except KeyError as exc:
-        raise EnableMetricsError(str(exc)) from exc
-
-
-def enable_replication(current_yaml: dict) -> None:
-    """Change the Synapse configuration to enable replication.
-
-    Args:
-        current_yaml: current configuration.
-
-    Raises:
-        WorkloadError: something went wrong enabling replication.
-    """
-    try:
-        resources = {"names": ["replication"]}
-        metric_listener = {
-            "port": 8034,
-            "type": "http",
-            "bind_addresses": ["::"],
-            "resources": [resources],
-        }
-        current_yaml["listeners"].extend([metric_listener])
-    except KeyError as exc:
-        raise WorkloadError(str(exc)) from exc
-
-
-def enable_forgotten_room_retention(current_yaml: dict) -> None:
-    """Change the Synapse configuration to enable forgotten_room_retention_period.
-
-    Args:
-        current_yaml: current configuration.
-    """
-    current_yaml["forgotten_room_retention_period"] = "28d"
-
-
-def enable_media_retention(current_yaml: dict) -> None:
-    """Change the Synapse configuration to enable media retention.
-
-    Args:
-        current_yaml: current configuration.
-    """
-    current_yaml["media_retention"] = {
-        "remote_media_lifetime": "14d",
-        "local_media_lifetime": "28d",
-    }
-
-
-def enable_stale_devices_deletion(current_yaml: dict) -> None:
-    """Change the Synapse configuration to delete stale devices.
-
-    Args:
-        current_yaml: current configuration.
-    """
-    current_yaml["delete_stale_devices_after"] = "1y"
-
-
-def disable_password_config(current_yaml: dict) -> None:
-    """Change the Synapse configuration to disable password config.
-
-    Args:
-        current_yaml: current configuration.
-    """
-    current_yaml["password_config"] = {"enabled": False}
-
-
-def disable_room_list_search(current_yaml: dict) -> None:
-    """Change the Synapse configuration to disable room_list_search.
-
-    Args:
-        current_yaml: current configuration.
-    """
-    current_yaml["enable_room_list_search"] = False
-
-
-def enable_serve_server_wellknown(current_yaml: dict) -> None:
-    """Change the Synapse configuration to enable server wellknown file.
-
-    Args:
-        current_yaml: current configuration.
-    """
-    current_yaml["serve_server_wellknown"] = True
-
-
-def enable_instance_map(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to instance_map config.
-
-    Args:
-        current_yaml: current configuration.
-        charm_state: Instance of CharmState.
-    """
-    current_yaml["instance_map"] = charm_state.instance_map_config
-
-
-def enable_stream_writers(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to stream_writers config.
-
-    Args:
-        current_yaml: current configuration.
-        charm_state: Instance of CharmState.
-    """
-    persisters = []
-    if charm_state.instance_map_config is not None:
-        persisters = [key for key in charm_state.instance_map_config.keys() if key != "main"]
-        persisters.sort()
-    if persisters is not None:
-        current_yaml["stream_writers"] = {"events": persisters}
-    else:
-        logger.error("Enable stream writers called but no persisters found. Verify peer relation.")
-
-
-def enable_federation_domain_whitelist(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to enable federation_domain_whitelist.
-
-    Args:
-        current_yaml: current configuration.
-        charm_state: Instance of CharmState.
-
-    Raises:
-        WorkloadError: something went wrong enabling configuration.
-    """
-    try:
-        federation_domain_whitelist = charm_state.synapse_config.federation_domain_whitelist
-        if federation_domain_whitelist is not None:
-            current_yaml["federation_domain_whitelist"] = _create_tuple_from_string_list(
-                federation_domain_whitelist
-            )
-    except KeyError as exc:
-        raise WorkloadError(str(exc)) from exc
-
-
-def enable_trusted_key_servers(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to set trusted_key_servers.
-
-    Args:
-        current_yaml: current configuration.
-        charm_state: Instance of CharmState.
-
-    Raises:
-        WorkloadError: something went wrong enabling configuration.
-    """
-    try:
-        trusted_key_servers = charm_state.synapse_config.trusted_key_servers
-        if trusted_key_servers is not None:
-            current_yaml["trusted_key_servers"] = tuple(
-                {"server_name": f"{item}"}
-                for item in _create_tuple_from_string_list(trusted_key_servers)
-            )
-    except KeyError as exc:
-        raise WorkloadError(str(exc)) from exc
-
-
-def enable_allow_public_rooms_over_federation(current_yaml: dict) -> None:
-    """Change the Synapse configuration to allow public rooms in federation.
-
-    Args:
-        current_yaml: current configuration.
-    """
-    current_yaml["allow_public_rooms_over_federation"] = True
-
-
-def _create_tuple_from_string_list(string_list: str) -> tuple[str, ...]:
-    """Format IP range whitelist.
-
-    Args:
-        string_list: comma separated list configuration.
-
-    Returns:
-        Tuple as expected by Synapse.
-    """
-    return tuple(item.strip() for item in string_list.split(","))
-
-
-def enable_ip_range_whitelist(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to enable ip_range_whitelist.
-
-    Args:
-        current_yaml: current configuration.
-        charm_state: Instance of CharmState.
-
-    Raises:
-        WorkloadError: something went wrong enabling configuration.
-    """
-    try:
-        ip_range_whitelist = charm_state.synapse_config.ip_range_whitelist
-        if ip_range_whitelist is None:
-            logger.warning("enable_ip_range_whitelist called but config is empty")
-            return
-        current_yaml["ip_range_whitelist"] = _create_tuple_from_string_list(ip_range_whitelist)
-    except KeyError as exc:
-        raise WorkloadError(str(exc)) from exc
-
-
-def _get_mjolnir_config(access_token: str, room_id: str) -> typing.Dict:
-    """Create config as expected by mjolnir.
-
-    Args:
-        access_token: access token to be used by the mjolnir bot.
-        room_id: management room id monitored by the Mjolnir.
-
-    Returns:
-        Mjolnir configuration
-    """
-    with open("templates/mjolnir_production.yaml", encoding="utf-8") as mjolnir_config_file:
-        config = yaml.safe_load(mjolnir_config_file)
-        config["homeserverUrl"] = SYNAPSE_URL
-        config["rawHomeserverUrl"] = SYNAPSE_URL
-        config["accessToken"] = access_token
-        config["managementRoom"] = room_id
-        return config
-
-
-def create_mjolnir_config(container: ops.Container, access_token: str, room_id: str) -> None:
-    """Create mjolnir configuration.
-
-    Args:
-        container: Container of the charm.
-        access_token: access token to be used by the Mjolnir.
-        room_id: management room id monitored by the Mjolnir.
-
-    Raises:
-        CreateMjolnirConfigError: something went wrong creating mjolnir config.
-    """
-    try:
-        config = _get_mjolnir_config(access_token, room_id)
-        container.push(MJOLNIR_CONFIG_PATH, yaml.safe_dump(config), make_dirs=True)
-    except ops.pebble.PathError as exc:
-        raise CreateMjolnirConfigError(str(exc)) from exc
-
-
-def _create_pysaml2_config(charm_state: CharmState) -> typing.Dict:
-    """Create config as expected by pysaml2.
-
-    Args:
-        charm_state: Instance of CharmState.
-
-    Returns:
-        Pysaml2 configuration.
-
-    Raises:
-        EnableSAMLError: if SAML configuration is not found.
-    """
-    if charm_state.saml_config is None:
-        raise EnableSAMLError(
-            "SAML Configuration not found. "
-            "Please verify the integration between SAML Integrator and Synapse."
-        )
-
-    saml_config = charm_state.saml_config
-    entity_id = (
-        charm_state.synapse_config.public_baseurl
-        if charm_state.synapse_config.public_baseurl is not None
-        else f"https://{charm_state.synapse_config.server_name}"
-    )
-    sp_config = {
-        "metadata": {
-            "remote": [
-                {
-                    "url": saml_config["metadata_url"],
-                },
-            ],
-        },
-        "allow_unknown_attributes": True,
-        "service": {
-            "sp": {
-                "entityId": entity_id,
-                "allow_unsolicited": True,
-            },
-        },
-    }
-    # login.staging.ubuntu.com and login.ubuntu.com
-    # dont send uid in SAMLResponse so this will map
-    # as expected
-    if "ubuntu.com" in saml_config["metadata_url"]:
-        sp_config["attribute_map_dir"] = "/usr/local/attributemaps"
-
-    return sp_config
-
-
-def enable_saml(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to enable SAML.
-
-    Args:
-        current_yaml: current configuration.
-        charm_state: Instance of CharmState.
-
-    Raises:
-        EnableSAMLError: something went wrong enabling SAML.
-    """
-    try:
-        if charm_state.synapse_config.public_baseurl is not None:
-            current_yaml["public_baseurl"] = charm_state.synapse_config.public_baseurl
-        # enable x_forwarded to pass expected headers
-        current_listeners = current_yaml["listeners"]
-        updated_listeners = [
-            {
-                **item,
-                "x_forwarded": (
-                    True
-                    if "x_forwarded" in item and not item["x_forwarded"]
-                    else item.get("x_forwarded", False)
-                ),
-            }
-            for item in current_listeners
-        ]
-        current_yaml["listeners"] = updated_listeners
-        current_yaml["saml2_enabled"] = True
-        current_yaml["saml2_config"] = {}
-        current_yaml["saml2_config"]["sp_config"] = _create_pysaml2_config(charm_state)
-        user_mapping_provider_config = {
-            "config": {
-                "mxid_source_attribute": "uid",
-                "grandfathered_mxid_source_attribute": "uid",
-                "mxid_mapping": "dotreplace",
-            },
-        }
-        current_yaml["saml2_config"]["user_mapping_provider"] = user_mapping_provider_config
-    except KeyError as exc:
-        raise EnableSAMLError(str(exc)) from exc
-
-
-def enable_smtp(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to enable SMTP.
-
-    Args:
-        current_yaml: current configuration.
-        charm_state: Instance of CharmState.
-
-    Raises:
-        EnableSMTPError: something went wrong enabling SMTP.
-    """
-    try:
-        current_yaml["email"] = {}
-        current_yaml["email"]["enable_notifs"] = charm_state.synapse_config.enable_email_notifs
-        current_yaml["email"]["notif_from"] = charm_state.synapse_config.notif_from
-
-        if charm_state.smtp_config is None:
-            raise EnableSMTPError(
-                "SMTP Configuration not found. "
-                "Please verify the integration between SMTP Integrator and Synapse."
-            )
-
-        smtp_config = charm_state.smtp_config
-        current_yaml["email"]["smtp_host"] = smtp_config["host"]
-        current_yaml["email"]["smtp_port"] = smtp_config["port"]
-        if charm_state.smtp_config["user"] is not None:
-            current_yaml["email"]["smtp_user"] = smtp_config["user"]
-        if charm_state.smtp_config["password"] is not None:
-            current_yaml["email"]["smtp_pass"] = smtp_config["password"]
-        current_yaml["email"]["enable_tls"] = smtp_config["enable_tls"]
-        current_yaml["email"]["force_tls"] = smtp_config["force_tls"]
-        current_yaml["email"]["require_transport_security"] = smtp_config[
-            "require_transport_security"
-        ]
-    except KeyError as exc:
-        raise EnableSMTPError(str(exc)) from exc
-
-
-def enable_media(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to enable S3.
-
-    Args:
-        current_yaml: Current Configuration.
-        charm_state: Instance of CharmState.
-
-    Raises:
-        WorkloadError: something went wrong enabling S3.
-    """
-    try:
-        if charm_state.media_config is None:
-            raise WorkloadError(
-                "Media Configuration not found. "
-                "Please verify the integration between Media and Synapse."
-            )
-        current_yaml["media_storage_providers"] = [
-            {
-                "module": "s3_storage_provider.S3StorageProviderBackend",
-                "store_local": True,
-                "store_remote": True,
-                "store_synchronous": True,
-                "config": {
-                    "bucket": charm_state.media_config["bucket"],
-                    "region_name": charm_state.media_config["region_name"],
-                    "endpoint_url": charm_state.media_config["endpoint_url"],
-                    "access_key_id": charm_state.media_config["access_key_id"],
-                    "secret_access_key": charm_state.media_config["secret_access_key"],
-                    "prefix": charm_state.media_config["prefix"],
-                },
-            },
-        ]
-    except KeyError as exc:
-        raise WorkloadError(str(exc)) from exc
-
-
-def enable_redis(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to enable Redis.
-
-    Args:
-        current_yaml: current configuration.
-        charm_state: Instance of CharmState.
-
-    Raises:
-        WorkloadError: something went wrong enabling SMTP.
-    """
-    try:
-        current_yaml["redis"] = {}
-
-        if charm_state.redis_config is None:
-            raise WorkloadError(
-                "Redis Configuration not found. "
-                "Please verify the integration between Redis and Synapse."
-            )
-
-        redis_config = charm_state.redis_config
-        current_yaml["redis"]["enabled"] = True
-        current_yaml["redis"]["host"] = redis_config["host"]
-        current_yaml["redis"]["port"] = redis_config["port"]
-    except KeyError as exc:
-        raise WorkloadError(str(exc)) from exc
-
-
-def enable_room_list_publication_rules(current_yaml: dict, charm_state: CharmState) -> None:
-    """Change the Synapse configuration to enable room_list_publication_rules.
-
-    This configuration is based on publish_rooms_allowlist charm configuration.
-    Once is set, a deny rule is added to prevent any other user to publish rooms.
-
-    Args:
-        current_yaml: current configuration.
-        charm_state: Instance of CharmState.
-
-    Raises:
-        WorkloadError: something went wrong enabling room_list_publication_rules.
-    """
-    room_list_publication_rules = []
-    # checking publish_rooms_allowlist to fix union-attr mypy error
-    publish_rooms_allowlist = charm_state.synapse_config.publish_rooms_allowlist
-    if publish_rooms_allowlist:
-        for user in publish_rooms_allowlist:
-            rule = {"user_id": user, "alias": "*", "room_id": "*", "action": "allow"}
-            room_list_publication_rules.append(rule)
-
-    if len(room_list_publication_rules) == 0:
-        raise WorkloadError("publish_rooms_allowlist has unexpected value. Please, verify it.")
-
-    last_rule = {"user_id": "*", "alias": "*", "room_id": "*", "action": "deny"}
-    room_list_publication_rules.append(last_rule)
-    current_yaml["room_list_publication_rules"] = room_list_publication_rules
-
-
 def reset_instance(container: ops.Container) -> None:
     """Erase data and config server_name.
 
@@ -756,41 +325,6 @@ def reset_instance(container: ops.Container) -> None:
                 "exception while erasing directory %s: %r", SYNAPSE_CONFIG_DIR, path_error
             )
             raise
-
-
-def get_environment(charm_state: CharmState) -> typing.Dict[str, str]:
-    """Generate a environment dictionary from the charm configurations.
-
-    Args:
-        charm_state: Instance of CharmState.
-
-    Returns:
-        A dictionary representing the Synapse environment variables.
-    """
-    environment = {
-        "SYNAPSE_CONFIG_DIR": SYNAPSE_CONFIG_DIR,
-        "SYNAPSE_CONFIG_PATH": SYNAPSE_CONFIG_PATH,
-        "SYNAPSE_DATA_DIR": SYNAPSE_DATA_DIR,
-        "SYNAPSE_REPORT_STATS": f"{charm_state.synapse_config.report_stats}",
-        "SYNAPSE_SERVER_NAME": f"{charm_state.synapse_config.server_name}",
-        # TLS disabled so the listener is HTTP. HTTPS will be handled by Traefik.
-        # TODO verify support to HTTPS backend before changing this  # pylint: disable=fixme
-        "SYNAPSE_NO_TLS": str(True),
-        "LD_PRELOAD": "/usr/lib/x86_64-linux-gnu/libjemalloc.so.2",
-    }
-    datasource = charm_state.datasource
-    if datasource is not None:
-        environment["POSTGRES_DB"] = datasource["db"]
-        environment["POSTGRES_HOST"] = datasource["host"]
-        environment["POSTGRES_PORT"] = datasource["port"]
-        environment["POSTGRES_USER"] = datasource["user"]
-        environment["POSTGRES_PASSWORD"] = datasource["password"]
-    for proxy_variable in ("http_proxy", "https_proxy", "no_proxy"):
-        proxy_value = getattr(charm_state.proxy, proxy_variable)
-        if proxy_value:
-            environment[proxy_variable] = str(proxy_value)
-            environment[proxy_variable.upper()] = str(proxy_value)
-    return environment
 
 
 def generate_nginx_config(container: ops.Container, main_unit_address: str) -> None:
@@ -854,3 +388,40 @@ def generate_worker_config(unit_number: str) -> dict:
         "worker_log_config": "/data/log.config",
     }
     return worker_config
+
+
+def _get_mjolnir_config(access_token: str, room_id: str) -> typing.Dict:
+    """Get config as expected by mjolnir.
+
+    Args:
+        access_token: access token to be used by the mjolnir bot.
+        room_id: management room id monitored by the Mjolnir.
+
+    Returns:
+        Mjolnir configuration
+    """
+    with open("templates/mjolnir_production.yaml", encoding="utf-8") as mjolnir_config_file:
+        config = yaml.safe_load(mjolnir_config_file)
+        config["homeserverUrl"] = SYNAPSE_URL
+        config["rawHomeserverUrl"] = SYNAPSE_URL
+        config["accessToken"] = access_token
+        config["managementRoom"] = room_id
+        return config
+
+
+def generate_mjolnir_config(container: ops.Container, access_token: str, room_id: str) -> None:
+    """Generate mjolnir configuration.
+
+    Args:
+        container: Container of the charm.
+        access_token: access token to be used by the Mjolnir.
+        room_id: management room id monitored by the Mjolnir.
+
+    Raises:
+        CreateMjolnirConfigError: something went wrong creating mjolnir config.
+    """
+    try:
+        config = _get_mjolnir_config(access_token, room_id)
+        container.push(MJOLNIR_CONFIG_PATH, yaml.safe_dump(config), make_dirs=True)
+    except ops.pebble.PathError as exc:
+        raise CreateMjolnirConfigError(str(exc)) from exc
