+++ tests/unit/test_charm.py
@@ -280,6 +280,7 @@ def test_enable_federation_domain_whitelist_is_called(
     harness.begin()
     monkeypatch.setattr(synapse, "execute_migrate_config", MagicMock())
     monkeypatch.setattr(synapse, "enable_metrics", MagicMock())
+    monkeypatch.setattr(synapse, "enable_rc_joins_remote_rate", MagicMock())
     monkeypatch.setattr(synapse, "enable_replication", MagicMock())
     monkeypatch.setattr(synapse, "enable_forgotten_room_retention", MagicMock())
     monkeypatch.setattr(synapse, "enable_serve_server_wellknown", MagicMock())
@@ -313,6 +314,7 @@ def test_disable_password_config_is_called(
     harness.begin()
     monkeypatch.setattr(synapse, "execute_migrate_config", MagicMock())
     monkeypatch.setattr(synapse, "enable_metrics", MagicMock())
+    monkeypatch.setattr(synapse, "enable_rc_joins_remote_rate", MagicMock())
     monkeypatch.setattr(synapse, "enable_replication", MagicMock())
     monkeypatch.setattr(synapse, "enable_forgotten_room_retention", MagicMock())
     monkeypatch.setattr(synapse, "enable_serve_server_wellknown", MagicMock())
