+++ tests/unit/test_synapse_workload.py
@@ -485,10 +485,10 @@ def test_get_mjolnir_config_success():
     assert config["managementRoom"] == room_id
 
 
-def test_create_mjolnir_config_success(monkeypatch: pytest.MonkeyPatch):
+def test_generate_mjolnir_config_success(monkeypatch: pytest.MonkeyPatch):
     """
     arrange: set container, access token and room id parameters.
-    act: call create_mjolnir_config.
+    act: call generate_mjolnir_config.
     assert: file is pushed as expected.
     """
     access_token = token_hex(16)
@@ -497,7 +497,7 @@ def test_create_mjolnir_config_success(monkeypatch: pytest.MonkeyPatch):
     container_mock = MagicMock()
     monkeypatch.setattr(container_mock, "push", push_mock)
 
-    synapse.create_mjolnir_config(
+    synapse.generate_mjolnir_config(
         container=container_mock, access_token=access_token, room_id=room_id
     )
 
@@ -757,3 +757,27 @@ def test_publish_rooms_allowlist_error(invalid_config):
         # Prevent mypy error:
         # Argument 1 to "SynapseConfig" has incompatible type "**dict[str, str]"; expected "bool"
         SynapseConfig(**synapse_with_notif_config)  # type: ignore[arg-type]
+
+
+def test_enable_rc_joins_remote_rate(
+    harness: Harness,
+    config_content: dict[str, typing.Any],
+):
+    """
+    arrange: set mock container with file.
+    act: update rc_joins_remote_rate config and call rc_joins_remote_rate.
+    assert: new configuration file is pushed and rc_joins_remote_rate is enabled.
+    """
+    config = config_content
+
+    harness.update_config({"rc_joins_remote_burst_count": 10, "rc_joins_remote_per_second": 0.2})
+    harness.begin()
+    synapse.enable_rc_joins_remote_rate(config, harness.charm.build_charm_state())
+
+    expected_config_content = {
+        "listeners": [
+            {"type": "http", "port": 8080, "bind_addresses": ["::"]},
+        ],
+        "rc_joins": {"remote": {"burst_count": 10, "per_second": 0.2}},
+    }
+    assert yaml.safe_dump(config) == yaml.safe_dump(expected_config_content)
